fix stats page distance display format error and user path tracking

+++ app/src/main/java/com/cs407/runtrackpro/StatsFragment.java
@@ -223,7 +223,7 @@ public class StatsFragment extends Fragment {
         }
 
         ArrayList<String> averages = new ArrayList<>();
-        averages.add(formatDistance((int) (totalDistance / size)));
+        averages.add(formatDistance(totalDistance / size));
         averages.add(formatTime((int) (totalTimeSeconds / size)));
         averages.add(formatPace((int) (totalPace / size)));
